third_party/WebKit/LayoutTests/fast/events/touch/touch-user-gesture-expected.txt

Target
Test user gesture behavior during touch events.
During a scroll, no touch events are user gestures
PASS openedPopup is false
PASS openedPopup is false
PASS openedPopup is false
During a drag that isn't a scroll, only touchend is a user gesture
PASS openedPopup is false
PASS openedPopup is false
PASS openedPopup is true
PASS successfullyParsed is true

TEST COMPLETE
